Added place holder sound files.  Attempting to debug NRE's that popped up when they did not previously exist.  Need to check my laptop settings because they don't exist on that machine.  Fixed AudioManager and IOData code to set volumes initially to 1 and also to change sound levels to floats vice ints.

+++ Blood Myth/Assets/Scripts/Systems/Data Persistence/IOData.cs
@@ -86,8 +86,8 @@ public class IOData
         }
 
     /*System Settings*/
-    private int _musiclevels;
-    public int MusicLevels
+    private float _musiclevels;
+    public float MusicLevels
     {
         get
         {
@@ -100,8 +100,8 @@ public class IOData
         }
     }
 
-    private int _sfxlevels;
-    public int SFXLevels
+    private float _sfxlevels;
+    public float SFXLevels
     {
         get
         {
